requirements.txt
openai==0.28
streamlit_authenticator

python-dotenv